mock/examples/onsnetwork.org/wp-scraper/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
html5lib==0.999999999
idna==2.6
numpy==1.13.1
python-dateutil==2.6.1
requests==2.18.4
six==1.11.0
urllib3==1.22
webencodings==0.5.1
